refactor(rocksdb): extract shared stats types to dedicated module
Move RocksDBTableStats and RocksDBStats to a shared rocksdb_types.rs
module that is used by both the real RocksDB provider and the stub
implementation. This eliminates code duplication and ensures type
consistency across platforms.

Amp-Thread-ID: https://ampcode.com/threads/T-019c226b-9f6a-74fc-a839-807adb52e6ab
