+++ crates/storage/provider/src/providers/rocksdb_stub.rs
@@ -15,32 +15,6 @@ use std::{path::Path, sync::Arc};
 /// Pending `RocksDB` batches type alias (stub - uses unit type).
 pub(crate) type PendingRocksDBBatches = Arc<Mutex<Vec<()>>>;
 
-/// Statistics for a single `RocksDB` table (column family) - stub.
-#[derive(Debug, Clone)]
-pub struct RocksDBTableStats {
-    /// Size of SST files on disk in bytes.
-    pub sst_size_bytes: u64,
-    /// Size of memtables in memory in bytes.
-    pub memtable_size_bytes: u64,
-    /// Name of the table/column family.
-    pub name: String,
-    /// Estimated number of keys in the table.
-    pub estimated_num_keys: u64,
-    /// Estimated size of live data in bytes (SST files + memtables).
-    pub estimated_size_bytes: u64,
-    /// Estimated bytes pending compaction (reclaimable space).
-    pub pending_compaction_bytes: u64,
-}
-
-/// Database-level statistics for `RocksDB` - stub.
-#[derive(Debug, Clone)]
-pub struct RocksDBStats {
-    /// Statistics for each table (column family).
-    pub tables: Vec<RocksDBTableStats>,
-    /// Total size of WAL (Write-Ahead Log) files in bytes.
-    pub wal_size_bytes: u64,
-}
-
 /// Context for `RocksDB` block writes (stub).
 #[derive(Debug, Clone)]
 #[allow(dead_code)]
